+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/Divider/Divider.tsx
@@ -1,0 +1,1 @@
+export const Divider = () => <div className="my-3 h-[1px] w-full bg-[#CECECE]"></div>;