+++ examples/kyb-app/src/common/utils/base64-to-file.ts
@@ -1,9 +1,0 @@
-export const base64ToFile = async (
-  base64: string,
-  fileName: string,
-  mimeType: string,
-): Promise<File> => {
-  return await fetch(base64)
-    .then(res => res.blob())
-    .then(blob => new File([blob], fileName, { type: mimeType }));
-};