+++ examples/kyb-app/src/components/organisms/DynamicUI/rule-engines/event.engine.ts
@@ -1,9 +1,0 @@
-import { RuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/rule-engine.abstract';
-
-export class EventEngine implements RuleEngine {
-  public ENGINE_NAME = 'event';
-
-  test() {
-    return { isValid: true, errors: [] };
-  }
-}