+++ examples/kyb-app/src/components/organisms/UIRenderer/elements/Divider/Divider.tsx
@@ -1,1 +1,0 @@
-export const Divider = () => <div className="bg-[#CECECE] w-full h-[1px] my-3"></div>;