+++ examples/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/components/LocalePicker/LocalePicker.tsx
@@ -1,8 +1,0 @@
-import { RJSVInputProps, TextInputAdapter } from '@ballerine/ui';
-import { languages } from './languages';
-
-export const LocalePicker = (props: RJSVInputProps) => {
-  props.schema.oneOf = languages;
-
-  return <TextInputAdapter {...(props as any)} />;
-};